requirements.txt
flask==3.0.0
flask-cors==4.0.0
email-validator==2.1.0
phonenumbers==8.13.26
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
loguru==0.7.2
